+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added support for dump files containing plain passwords
 - Added "clean" parameter for the classify job, so you can automatically clean previous results
 - Moved the features file up one level for better reading
+- Added shortcut to open the current file while training the system
 
 ## 0.2.0 [Unreleased]
 ### Added